added userRole repository, permission repository, PermissionsController and UserController

+++ ShopApp.Application/Services/RolePermission/RolePermissionService.cs
@@ -12,31 +12,29 @@ public class RolePermissionService : IRolePermissionService
 
     private readonly IPermissionActionRepository _permissionActionRepository;
     private readonly IRoleRepository _roleRepository;
-    private readonly IMapper _mapper;
 
-    public RolePermissionService(IRolePermissionRepository rolePermissionRepository, 
+    public RolePermissionService(IRolePermissionRepository rolePermissionRepository,
     IPermissionActionRepository permissionActionRepository,
-    IRoleRepository roleRepository, IMapper mapper)
+    IRoleRepository roleRepository)
     {
         _rolePermissionRepository = rolePermissionRepository;
         _permissionActionRepository = permissionActionRepository;
         _roleRepository = roleRepository;
-        _mapper = mapper;
-    }
-    public async Task AssignPermissionsToRole(PermissionsToRoleDto assignPermissionsToRoleDto)
-    {
 
-        var rolePermissions = assignPermissionsToRoleDto.PermissionActionIds.Select(
+    }
+    public async Task AssignPermissionsToRole(Guid RoleId, Guid[] PermissionActionIds)
+    {
+        var rolePermissions = PermissionActionIds.Select(
             x => new RolePermission
             {
                 PermissionActionId = x,
-                RoleId = assignPermissionsToRoleDto.RoleId
+                RoleId = RoleId
             }
         ).ToList();
 
         foreach (var rolePermission in rolePermissions)
         {
-            if(!await _roleRepository.IsExists(rolePermission.RoleId ?? Guid.Empty) && 
+            if (!await _roleRepository.IsExists(rolePermission.RoleId ?? Guid.Empty) &&
             !await _permissionActionRepository.IsExists(rolePermission.PermissionActionId ?? Guid.Empty))
             {
                 throw new Exception("Such role or permission doesn't exist");
@@ -56,29 +54,33 @@ public class RolePermissionService : IRolePermissionService
 
     }
 
-    public async Task UnAssignPermissionsFromRole(PermissionsToRoleDto assignPermissionsToRoleDto)
+    public async Task UnAssignPermissionsFromRole(Guid RoleId, Guid[] PermissionActionIds)
     {
-          var rolePermissions = assignPermissionsToRoleDto.PermissionActionIds.Select(
-            x => new RolePermission
-            {
-                PermissionActionId = x,
-                RoleId = assignPermissionsToRoleDto.RoleId
-            }
-        ).ToList();
+        var rolePermissions = PermissionActionIds.Select(
+          x => new RolePermission
+          {
+              PermissionActionId = x,
+              RoleId = RoleId
+          }
+          ).ToList();
 
         foreach (var rolePermission in rolePermissions)
         {
+            if (!await _roleRepository.IsExists(rolePermission.RoleId ?? Guid.Empty) &&
+            !await _permissionActionRepository.IsExists(rolePermission.PermissionActionId ?? Guid.Empty))
+            {
+                throw new Exception("Such role or permission doesn't exist");
+            }
 
-        
             if (!await _rolePermissionRepository.IsRoleExistsWithSuchPermission(
                  rolePermission.RoleId ?? Guid.Empty,
                  rolePermission.PermissionActionId ?? Guid.Empty))
             {
-                    throw new Exception("There are no such permission that assigned to such role");
+                throw new Exception("There are no such permission that assigned to such role");
             }
             await _rolePermissionRepository.DeleteAsync(rolePermission.Id);
         }
-        
+
         await _rolePermissionRepository.SaveAsync();
 
     }
